# Documentation/swagger.yaml
openapi: 3.0.0
info:
  title: StudyMove API
  description: Documentation de l'API StudyMove
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Serveur local de développement
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/v1/events/create:
    post:
      summary: "Créer un événement"
      description: "Crée un nouvel événement avec des détails fournis par l'utilisateur."
      operationId: "createEvent"
      tags:
        - "Events"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: "Titre de l'événement"
                  example: "Concert de Musique"
                subtitle:
                  type: string
                  description: "Sous-titre de l'événement"
                  example: "Une soirée de musique en live"
                start_date:
                  type: string
                  format: date-time
                  description: "Date et heure de début de l'événement"
                  example: "2024-12-20T18:00:00Z"
                end_date:
                  type: string
                  format: date-time
                  description: "Date et heure de fin de l'événement"
                  example: "2024-12-20T23:00:00Z"
                start_time:
                  type: string
                  format: date-time
                  description: "Heure de début pour affichage spécifique"
                  example: "2024-12-20T18:00:00Z"
                end_time:
                  type: string
                  format: date-time
                  description: "Heure de fin pour affichage spécifique"
                  example: "2024-12-20T23:00:00Z"
                is_online:
                  type: boolean
                  description: "Indique si l'événement est en ligne"
                  example: false
                is_visible:
                  type: boolean
                  description: "Indique si l'événement est visible pour tous"
                  example: true
                use_studibox:
                  type: boolean
                  description: "Utilise la billetterie Studibox si vrai"
                  example: true
                ticket_price:
                  type: number
                  format: float
                  description: "Prix d'un billet standard"
                  example: 50.0
                ticket_stock:
                  type: integer
                  description: "Nombre total de billets disponibles"
                  example: 200
                location:
                  type: object
                  description: "Détails de l'emplacement de l'événement"
                  properties:
                    address:
                      type: string
                      example: "123 Rue de la Musique"
                    city:
                      type: string
                      example: "Paris"
                    postalCode:
                      type: string
                      example: "75001"
                    region:
                      type: string
                      example: "Île-de-France"
                    country:
                      type: string
                      example: "France"
                categories:
                  type: array
                  items:
                    type: string
                  description: "Liste des catégories pour l'événement"
                  example: ["Concert", "Live Music"]
                tags:
                  type: array
                  items:
                    type: string
                  description: "Liste des tags pour l'événement"
                  example: ["Pop", "Jazz", "Night Event"]
                options:
                  type: array
                  description: "Options additionnelles avec tarifs et stock"
                  items:
                    type: object
                    properties:
                      title:
                        type: string
                        example: "VIP Access"
                      description:
                        type: string
                        example: "Accès au lounge VIP et places premium"
                      price:
                        type: number
                        format: float
                        example: 120.0
                      stock:
                        type: integer
                        example: 30
                tarifs:
                  type: array
                  description: "Différents types de billets avec tarifs et stock"
                  items:
                    type: object
                    properties:
                      title:
                        type: string
                        example: "Entrée Standard"
                      description:
                        type: string
                        example: "Accès standard au concert"
                      price:
                        type: number
                        format: float
                        example: 50.0
                      stock:
                        type: integer
                        example: 150
                descriptions:
                  type: array
                  description: "Informations détaillées sur l'événement"
                  items:
                    type: object
                    properties:
                      title:
                        type: string
                        example: "Présentation de l'événement"
                      description:
                        type: string
                        example: "Un concert de musique avec les meilleurs artistes de l'année."
                external_ticketing_url:
                  type: string
                  description: "URL pour la billetterie externe si `use_studibox` est `false`"
                  example: "https://external.tickets.com/event123"
                video_url:
                  type: string
                  description: "URL de la vidéo de présentation de l'événement"
                  example: "https://video.link/concert"
                images:
                  type: array
                  items:
                    type: string
                  description: "URLs des images associées à l'événement"
                  example:
                    [
                      "https://image.link/img1.jpg",
                      "https://image.link/img2.jpg",
                    ]
      responses:
        "201":
          description: "Événement créé avec succès"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  eventID:
                    type: integer
                    description: "ID de l'événement créé"
                    example: 1
        "400":
          description: "Requête invalide"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid request payload"
        "401":
          description: "Non autorisé, token manquant ou invalide"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Unauthorized"
        "500":
          description: "Erreur interne du serveur"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Failed to create event"
...